Add clear cookies button
We were experiencing an issue with login. The user would click login,
and then get redirected to the home page, without being logged in.

I never figured out why this was occurring, but I did work out that
clearing cookies starting with wordpress_ always fixed the problem in my
testing.

For this reason, I added a button labelled "Clear cookies" in the login
page for users to use.